requirements.txt
# Political Polarization Experiment - Python Dependencies

# Core
numpy>=1.24.0
torch>=2.0.0

# LLM Inference
vllm>=0.4.0
openai>=1.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
